artifacts/count_people/requirements.txt
awscrt
awsiotsdk
gluoncv 
labels
mxnet